# docs/openapi.yaml
openapi: 3.1.0
info:
  title: eTradeBazaar API
  version: 1.0.0
  description: |
    REST API for the eTradeBazaar multi-tenant marketplace platform.

    The platform operates across three actor types: platform administrators,
    sellers (multi-member tenants), and customers. Authentication is JWT-based.
    All protected endpoints require a Bearer token in the Authorization header.

    Base URL: https://api.etradebazaar.com
servers:
  - url: https://api.etradebazaar.com
    description: Production
  - url: https://staging-api.etradebazaar.com
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Obtain a token via POST /api/auth/login. Pass it as:
        Authorization: Bearer <token>
...
  schemas:
...
    MessageEnvelope:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Resource deleted.
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Resource not found.
...
    User:
      type: object
      properties:
        id:
          type: string
          example: clxyz1234abcd
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          nullable: true
          example: Rahul Mehta
        isActive:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            example:
              error: No token provided
    Forbidden:
      description: Authenticated but insufficient role or account not approved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: Insufficient platform permissions
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
              retryAfter:
                type: integer
          example:
            success: false
            error: Too many requests - please try again later
            retryAfter: 45
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      description: |
        Authenticates a user with email and password. Returns a short-lived access token
        and a long-lived refresh token. Both sellers and platform admins use this endpoint.

        Rate limit: 10 requests per minute per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: seller@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: secret123
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  accessToken:
                    type: string
                    description: JWT access token. Short-lived.
                  refreshToken:
                    type: string
                    description: JWT refresh token. Valid for 7 days.
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                        nullable: true
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                error: Invalid credentials
        "403":
          description: Account is disabled.
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      description: |
        Invalidates the current access token by adding its JTI to a Redis blacklist.
        The token TTL is respected so the blacklist entry expires naturally.
      responses:
        "200":
          description: Logged out successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
              example:
                success: true
                message: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      description: |
        Issues a new access token and rotates the refresh token. Pass the refresh token
        in the request body or as the `refreshToken` cookie.

        Rate limit: 10 requests per minute per IP.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Required if not sent as a cookie.
      responses:
        "200":
          description: Tokens refreshed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        "401":
          description: Missing or invalid refresh token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: Invalid refresh token
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      description: |
        Returns the authenticated user's profile. If the user is a seller, the response
        includes the seller's current approval status.
      responses:
        "200":
          description: Current user profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - type: object
                        properties:
                          sellerId:
                            type: string
                            nullable: true
                          role:
                            type: string
                            example: seller
                          status:
                            type: string
                            nullable: true
                            description: Present when the user is a seller.
                          isApproved:
                            type: boolean
                            nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
...